requirements.txt
apache-airflow
sklearn
apache-airflow-providers-cncf-kubernetes
